refactor(#4): create src/ folder with Clean Arch directories
