+++ app/(tabs)/index.tsx
@@ -1,10 +1,10 @@
 import { Image } from 'expo-image';
 import { Platform, StyleSheet } from 'react-native';
 
-import { HelloWave } from '@/components/hello-wave';
-import ParallaxScrollView from '@/components/parallax-scroll-view';
-import { ThemedText } from '@/components/themed-text';
-import { ThemedView } from '@/components/themed-view';
+import { HelloWave } from '@/src/presentation/components/hello-wave';
+import ParallaxScrollView from '@/src/presentation/components/parallax-scroll-view';
+import { ThemedText } from '@/src/presentation/components/themed-text';
+import { ThemedView } from '@/src/presentation/components/themed-view';
 import { Link } from 'expo-router';
 
 export default function HomeScreen() {
