+++ app/_layout.tsx
@@ -3,7 +3,7 @@ import { Stack } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import 'react-native-reanimated';
 
-import { useColorScheme } from '@/hooks/use-color-scheme';
+import { useColorScheme } from '@/src/presentation/hooks/use-color-scheme';
 
 export const unstable_settings = {
   anchor: '(tabs)',
