+++ app/modal.tsx
@@ -1,15 +1,15 @@
 import { Link } from 'expo-router';
 import { StyleSheet } from 'react-native';
 
-import { ThemedText } from '@/components/themed-text';
-import { ThemedView } from '@/components/themed-view';
+import { ThemedText } from '@/src/presentation/components/themed-text';
+import { ThemedView } from '@/src/presentation/components/themed-view';
 
 export default function ModalScreen() {
   return (
     <ThemedView style={styles.container}>
-      <ThemedText type="title">This is a modal</ThemedText>
-      <Link href="/" dismissTo style={styles.link}>
-        <ThemedText type="link">Go to home screen</ThemedText>
+      <ThemedText type='title'>This is a modal</ThemedText>
+      <Link href='/' dismissTo style={styles.link}>
+        <ThemedText type='link'>Go to home screen</ThemedText>
       </Link>
     </ThemedView>
   );
