+++ src/presentation/components/external-link.tsx
@@ -7,7 +7,7 @@ type Props = Omit<ComponentProps<typeof Link>, 'href'> & { href: Href & string }
 export function ExternalLink({ href, ...rest }: Props) {
   return (
     <Link
-      target="_blank"
+      target='_blank'
       {...rest}
       href={href}
       onPress={async (event) => {
