+++ src/presentation/components/haptic-tab.tsx
