+++ src/presentation/components/hello-wave.tsx
@@ -12,7 +12,8 @@ export function HelloWave() {
         },
         animationIterationCount: 4,
         animationDuration: '300ms',
-      }}>
+      }}
+    >
       👋
     </Animated.Text>
   );
