+++ src/presentation/components/ui/icon-symbol.ios.tsx
